+++ bin/lib/getDependentStoryFiles.js
@@ -1,3 +1,6 @@
+import path from 'path';
+
+import { getWorkingDir } from '../git/git';
 import bailFile from '../ui/messages/warnings/bailFile';
 
 // Bail whenever one of these was changed
@@ -9,12 +12,18 @@ const EXTERNALS = [/\/node_modules\//, /\/webpack\/runtime\//, /^\(webpack\)/];
 const isGlobal = (name) => GLOBALS.some((re) => re.test(name));
 const isUserCode = ({ name, moduleName }) => !EXTERNALS.some((re) => re.test(name || moduleName));
 
-export function getDependentStoryFiles(ctx, stats, changedFiles) {
-  const { configDir = './.storybook', staticDir = [] } = ctx.storybook || {};
+const posixRelativePath = (dir) => `./${dir.split(path.sep).filter(Boolean).join(path.posix.sep)}`;
 
-  // TODO deal with Windows path separator
-  const storybookDir = configDir.startsWith('./') ? configDir : `./${configDir}`;
-  const staticDirs = staticDir.map((dir) => (dir.startsWith('./') ? dir : `./${dir}`));
+export async function getDependentStoryFiles(ctx, stats, changedFiles) {
+  const { configDir = '.storybook', staticDir = [] } = ctx.storybook || {};
+  const workingDir = await getWorkingDir();
+  const workingDirRegExp = workingDir && new RegExp(`^./${workingDir}`);
+
+  // Replace Windows-style backslash path separators with POSIX-style forward slashes, because the
+  // Webpack stats use forward slashes in the `name` and `moduleName` fields. Note `changedFiles`
+  // already contains forward slashes, because that's what git yields even on Windows.
+  const storybookDir = posixRelativePath(configDir);
+  const staticDirs = staticDir.map(posixRelativePath);
 
   // NOTE: this only works with `main:stories` -- if stories are imported from files in `.storybook/preview.js`
   // we'll need a different approach to figure out CSF files (maybe the user should pass a glob?).
@@ -68,7 +77,10 @@ export function getDependentStoryFiles(ctx, stats, changedFiles) {
     }
   }
 
-  changedFiles.forEach(traceName);
+  changedFiles.forEach((gitFilePath) => {
+    const webpackFilePath = workingDir ? gitFilePath.replace(workingDirRegExp, '.') : gitFilePath;
+    traceName(webpackFilePath);
+  });
   while (toCheck.length > 0) {
     const id = toCheck.pop();
     checkedIds[id] = true;