+++ bin/tasks/gitInfo.test.js
@@ -1,18 +1,27 @@
 import { getCommitAndBranch } from '../git/getCommitAndBranch';
-import { getParentCommits, getSlug, getVersion } from '../git/git';
+import {
+  getBaselineBuilds,
+  getChangedFiles,
+  getParentCommits,
+  getSlug,
+  getVersion,
+} from '../git/git';
 import { setGitInfo } from './gitInfo';
 
 jest.mock('../git/getCommitAndBranch');
 jest.mock('../git/git');
 
-const log = { debug: jest.fn() };
+const log = { warn: jest.fn(), debug: jest.fn() };
+
+beforeEach(() => {
+  getCommitAndBranch.mockResolvedValue({ commit: '123asdf', branch: 'something' });
+  getParentCommits.mockResolvedValue(['asd2344']);
+  getVersion.mockResolvedValue('Git v1.0.0');
+  getSlug.mockResolvedValue('user/repo');
+});
 
 describe('setGitInfo', () => {
   it('sets the git info on context', async () => {
-    getCommitAndBranch.mockResolvedValue({ commit: '123asdf', branch: 'something' });
-    getParentCommits.mockResolvedValue(['asd2344']);
-    getVersion.mockResolvedValue('Git v1.0.0');
-    getSlug.mockResolvedValue('user/repo');
     const ctx = { log, options: {} };
     await setGitInfo(ctx, {});
     expect(ctx.git).toMatchObject({
@@ -25,18 +34,24 @@ describe('setGitInfo', () => {
   });
 
   it('supports overriding the owner name in the slug', async () => {
-    getCommitAndBranch.mockResolvedValue({ commit: '123asdf', branch: 'something' });
-    getParentCommits.mockResolvedValue(['asd2344']);
-    getVersion.mockResolvedValue('Git v1.0.0');
-    getSlug.mockResolvedValue('user/repo');
     const ctx = { log, options: { ownerName: 'org' } };
     await setGitInfo(ctx, {});
-    expect(ctx.git).toMatchObject({
-      commit: '123asdf',
-      branch: 'something',
-      parentCommits: ['asd2344'],
-      version: 'Git v1.0.0',
-      slug: 'org/repo',
-    });
+    expect(ctx.git).toMatchObject({ slug: 'org/repo' });
+  });
+
+  it('sets changedFiles', async () => {
+    getBaselineBuilds.mockResolvedValue([{ commit: '012qwes' }]);
+    getChangedFiles.mockResolvedValue(['styles/main.scss', 'lib/utils.js']);
+    const ctx = { log, options: { onlyChanged: true } };
+    await setGitInfo(ctx, {});
+    expect(ctx.git).toMatchObject({ changedFiles: ['./styles/main.scss', './lib/utils.js'] });
+  });
+
+  it('drops changedFiles when matching --externals', async () => {
+    getBaselineBuilds.mockResolvedValue([{ commit: '012qwes' }]);
+    getChangedFiles.mockResolvedValue(['styles/main.scss', 'lib/utils.js']);
+    const ctx = { log, options: { onlyChanged: true, externals: ['*.scss'] } };
+    await setGitInfo(ctx, {});
+    expect(ctx.git).not.toHaveProperty('changedFiles');
   });
 });